backend/requirements.txt
fastapi
uvicorn
python-multipart
rembg[cpu]
pillow
opencv-python-headless
numpy
httpx
